refactor : rename hook to useOnClickOutside and use single parameter to handle callback
